backend1/stress_service/requirements.txt
Flask
gunicorn
numpy
tensorflow
scikit-learn
joblib
python-dotenv